Split integration_test to more modules
