+++ crates/cargo-bsp/tests/common/mod.rs
@@ -1,1 +1,5 @@
-pub mod client;
+mod client;
+pub use client::*;
+
+mod server_utils;
+pub use server_utils::*;
